requirements.txt
httpx>=0.28.1
python-dotenv>=1.2.2
requests>=2.31.0
selenium>=4.20.0